feat(model-provider): update tests and documentation after rebase
- Fix test expectations to include headers field
- Consolidate duplicate test files
- Update README with Claude headers documentation
- Add comprehensive integration tests for Claude headers
