+++ multimodal/tarko/model-provider/tests/model-resolver.test.ts
@@ -17,6 +17,7 @@ describe('resolveModel', () => {
       displayName: undefined,
       baseURL: undefined,
       apiKey: undefined,
+      headers: {},
       baseProvider: 'openai',
     });
   });
@@ -38,6 +39,9 @@ describe('resolveModel', () => {
       displayName: 'Claude 3.5 Sonnet',
       baseURL: 'https://api.anthropic.com',
       apiKey: 'test-key',
+      headers: {
+        'anthropic-beta': 'fine-grained-tool-streaming-2025-05-14,token-efficient-tools-2025-02-19',
+      },
       baseProvider: 'anthropic',
     });
   });
@@ -57,6 +61,7 @@ describe('resolveModel', () => {
       displayName: undefined,
       baseURL: undefined,
       apiKey: 'original-key',
+      headers: {},
       baseProvider: 'anthropic',
     });
   });
@@ -70,6 +75,7 @@ describe('resolveModel', () => {
       displayName: undefined,
       baseURL: 'http://127.0.0.1:11434/v1',
       apiKey: 'ollama',
+      headers: {},
       baseProvider: 'openai',
     });
   });
@@ -90,6 +96,7 @@ describe('resolveModel', () => {
       displayName: undefined,
       baseURL: 'http://custom-host:8080/v1',
       apiKey: 'custom-key',
+      headers: {},
       baseProvider: 'openai',
     });
   });
@@ -103,6 +110,7 @@ describe('resolveModel', () => {
       displayName: undefined,
       baseURL: 'https://ark.cn-beijing.volces.com/api/v3',
       apiKey: undefined,
+      headers: {},
       baseProvider: 'openai',
     });
   });
@@ -122,6 +130,7 @@ describe('resolveModel', () => {
       displayName: undefined,
       baseURL: 'https://api.deepseek.com/v1',
       apiKey: 'deepseek-key',
+      headers: {},
       baseProvider: 'openai',
     });
   });
